+++ instrumentation/opentelemetry_finch/lib/opentelemetry_finch.ex
@@ -21,31 +21,34 @@ defmodule OpentelemetryFinch do
   require OpenTelemetry.Tracer
 
   @typedoc "Setup options"
-  @type opts :: []
+  @type opts :: [
+          {:req_headers_to_span_attributes, String.t()}
+          | {:resp_headers_to_span_attributes, String.t()}
+        ]
 
   @doc """
   Initializes and configures the telemetry handlers.
   """
   @spec setup(opts()) :: :ok
-  def setup(_opts \\ []) do
+  def setup(opts \\ []) do
     :telemetry.attach(
       {__MODULE__, :request_stop},
       [:finch, :request, :stop],
       &__MODULE__.handle_request_stop/4,
-      %{}
+      parse_opts(opts)
     )
   end
 
   @doc false
-  def handle_request_stop(_event, measurements, meta, _config) do
+  def handle_request_stop(_event, measurements, meta, config) do
     duration = measurements.duration
     end_time = :opentelemetry.timestamp()
     start_time = end_time - duration
 
-    status =
+    {response, status} =
       case meta.result do
-        {:ok, response} -> response.status
-        _ -> 0
+        {:ok, response} -> {response, response.status}
+        _ -> {%{}, 0}
       end
 
     url = build_url(meta.request.scheme, meta.request.host, meta.request.port, meta.request.path)
@@ -59,6 +62,25 @@ defmodule OpentelemetryFinch do
       Trace.http_method() => meta.request.method,
       Trace.http_status_code() => status
     }
+
+    req_headers_attributes =
+      :opentelemetry_instrumentation_http.extract_headers_attributes(
+        :request,
+        meta.request.headers,
+        Map.get(config, :req_headers_to_span_attributes, [])
+      )
+
+    resp_headers_attributes =
+      :opentelemetry_instrumentation_http.extract_headers_attributes(
+        :response,
+        Map.get(response, :headers, []),
+        Map.get(config, :resp_headers_to_span_attributes, [])
+      )
+
+    attributes =
+      attributes
+      |> Map.merge(req_headers_attributes)
+      |> Map.merge(resp_headers_attributes)
 
     s =
       OpenTelemetry.Tracer.start_span("HTTP #{meta.request.method}", %{
@@ -81,4 +103,15 @@ defmodule OpentelemetryFinch do
 
   defp format_error(%{__exception__: true} = exception), do: Exception.message(exception)
   defp format_error(reason), do: inspect(reason)
+
+  defp parse_opts(opts) do
+    opts
+    |> Map.new()
+    |> Map.update(:req_headers_to_span_attributes, [], &parse_headers_to_span_attributes/1)
+    |> Map.update(:resp_headers_to_span_attributes, [], &parse_headers_to_span_attributes/1)
+  end
+
+  defp parse_headers_to_span_attributes(headers) do
+    Enum.map(headers, &:opentelemetry_instrumentation_http.normalize_header_name/1)
+  end
 end