+++ instrumentation/opentelemetry_finch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+* New options (`req_headers_to_span_attributes`  and `resp_headers_to_span_attributes`) to automatically add request/response headers as attributes 
+
 ## 0.2.0
 
 * Span attributes update