+++ instrumentation/opentelemetry_httpoison/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+* New options (`req_headers_to_span_attributes`  and `resp_headers_to_span_attributes`) to automatically add request/response headers as attributes 
+
 ## 1.3.0
 
 ### Changed